[+] working make file for linux

+++ src/The-Powder-Toy-3D.cpp
@@ -14,15 +14,12 @@ using namespace sf;
 
 int main()
 {
-
-
-
     lld_t* lld = lld_init();
 
-    lld_insert(lld, (int)lld->data, (void *)"hello ");
-    lld_insert(lld, (int)lld->data, (void*)"how ");
-    lld_insert(lld, (int)lld->data, (void*)"are ");
-    lld_insert(lld, (int)lld->data, (void*)"you? ");
+    lld_insert(lld, (int)(i64)lld->data, (void *)"hello ");
+    lld_insert(lld, (int)(i64)lld->data, (void*)"how ");
+    lld_insert(lld, (int)(i64)lld->data, (void*)"are ");
+    lld_insert(lld, (int)(i64)lld->data, (void*)"you? ");
 
     while (lld->data) {
         char* str = (char*)lld_pop(lld, 0);
